chore(repo): General maintenance and fixes
Includes a variety of small fixes, code improvements, and documentation updates across the codebase.

- **README**: The installation guide in `README.md` has been updated to remove the incorrect manual compilation command. The `make` command is now the only recommended way to build the compiler.
- **Build**: Fixed a number of compiler warnings related to signed/unsigned comparisons and pessimizating moves.
- **Includes**: Corrected include paths and added missing headers to resolve build issues.
- **Types**: Added the `!=` operator for `DataType` for more convenient type comparisons.

+++ src/backend/CodeGen.cpp
@@ -861,7 +861,7 @@ CodeGen::passParams(const std::vector<std::shared_ptr<TACKY::Val>> &params, bool
     }
 
     // pass params in INTEGER regsiters
-    for (int i = 0; i < intRegParams.size(); i++)
+    for (size_t i = 0; i < intRegParams.size(); i++)
     {
         auto [paramType, param] = intRegParams[i];
         auto r = remainingIntRegs[i];
@@ -877,7 +877,7 @@ CodeGen::passParams(const std::vector<std::shared_ptr<TACKY::Val>> &params, bool
     }
 
     // pass params in DOUBLE regsiters
-    for (int i = 0; i < dblRegParams.size(); i++)
+    for (size_t i = 0; i < dblRegParams.size(); i++)
     {
         auto param = dblRegParams[i];
         auto r = DBL_PARAM_PASSING_REGS[i];
@@ -886,7 +886,7 @@ CodeGen::passParams(const std::vector<std::shared_ptr<TACKY::Val>> &params, bool
 
     // pass params on the stack
     // first param passed on stack has index 0 and is passed at Stack(16)
-    for (int i = 0; i < stackParams.size(); i++)
+    for (size_t i = 0; i < stackParams.size(); i++)
     {
         auto [paramType, param] = stackParams[i];
         auto stack = std::make_shared<Assembly::Memory>(std::make_shared<Assembly::Reg>(Assembly::RegName::BP), 16 + 8 * i);
